feat(app): 메인 페이지 라우팅 및 UI/UX 대규모 리팩토링
- **라우팅 구조 변경**:
  - 기존의 정보 제공용 랜딩 페이지를 `/about`으로 이동
  - 핵심 기능인 '번역 캔버스'를 메인 루트(`/`)로 변경
  - 인증되지 않은 사용자는 `/about`으로 리디렉션되도록 `middleware.ts` 업데이트

- **UI/UX 개선**:
  - '새 번역'과 '번역 결과' 뷰를 분리하여 전체 화면 공간을 활용하도록 개선. 이를 통해 기존의 복잡한 스크롤 문제를 근본적으로 해결
  - 메인 페이지와 About 페이지 간의 상호 이동을 위한 네비게이션 버튼 추가

- **버그 수정**:
  - About 페이지에서 로그인된 사용자를 강제로 리디렉션하던 문제 수정
  - About 페이지의 컴포넌트 경로 문제 해결

- **코드 정리**:
  - 라우팅 변경에 따라 `/canvas` 관련 파일 및 디렉토리 삭제
  - 코드 내의 모든 `/canvas` 경로를 `/`로 업데이트
