+++ frontend/src/app/(auth)/sign-up/[[...sign-up]]/page.tsx
@@ -3,7 +3,7 @@ import { SignUp } from "@clerk/nextjs";
 export default function Page() {
   return (
     <SignUp 
-      fallbackRedirectUrl="/canvas"
+      fallbackRedirectUrl="/"
       signInUrl="/sign-in"
     />
   );
