--- frontend/src/app/canvas/layout.tsx
@@ -1,13 +1,0 @@
-import { ReactNode } from 'react';
-
-export default function CanvasLayout({
-  children,
-}: {
-  children: ReactNode;
-}) {
-  return (
-    <div className="min-h-screen bg-gray-50">
-      {children}
-    </div>
-  );
-}
